+++ docs/source/_examples/single_plot_2D_imshow.py
@@ -13,7 +13,7 @@ image, extents = ap.load_2d_from_root(
     output_format="imshow")
 
 # format figure
-plt.rcParams["image.cmap"] = "lmf2root"
+plt.rcParams["image.cmap"] = "atom"
 plt.rcParams["image.aspect"] = "auto"
 plt.rcParams["image.interpolation"] = "none"